feat: implement automatic color generation for charts and add dataset color handling functions
