+++ constants/Colors.ts
@@ -47,3 +47,51 @@ export const semanticColors = {
 } as const;
 
 export default colors;
+
+// Paleta de colores para gráficos - Generación automática
+export const chartColorPalette = [
+  '#1C366B', // Azul primario institucional
+  '#36A2EB', // Azul claro
+  '#FF6384', // Rosa/Rojo
+  '#FFCE56', // Amarillo
+  '#4BC0C0', // Turquesa
+  '#9966FF', // Púrpura
+  '#FF9F40', // Naranja
+  '#FF6B6B', // Rojo coral
+  '#4ECDC4', // Verde agua
+  '#45B7D1', // Azul cielo
+  '#96CEB4', // Verde menta
+  '#FECA57', // Amarillo dorado
+  '#FF9FF3', // Rosa claro
+  '#54A0FF', // Azul brillante
+  '#5F27CD', // Púrpura oscuro
+  '#00D2D3', // Cian
+  '#FF9F43', // Naranja claro
+  '#10AC84', // Verde esmeralda
+  '#EE5A24', // Naranja rojizo
+  '#0984E3'  // Azul intenso
+];
+
+// Función para generar colores automáticamente según la cantidad de datos
+export const generateChartColors = (count: number): string[] => {
+  const colors: string[] = [];
+  for (let i = 0; i < count; i++) {
+    colors.push(chartColorPalette[i % chartColorPalette.length]);
+  }
+  return colors;
+};
+
+// Función para generar colores con opacidad para datasets
+export const generateDatasetColors = (count: number) => {
+  const baseColors = generateChartColors(count);
+  return baseColors.map(color => ({
+    color: (opacity = 1) => {
+      // Convertir hex a rgba
+      const r = parseInt(color.slice(1, 3), 16);
+      const g = parseInt(color.slice(3, 5), 16);
+      const b = parseInt(color.slice(5, 7), 16);
+      return `rgba(${r}, ${g}, ${b}, ${opacity})`;
+    },
+    strokeWidth: 3
+  }));
+};
